+++ src/generators/http.py
@@ -24,8 +24,24 @@ class HTTPGenerator:
         self.output_dir = output_dir
         self.results: dict[str, int] = defaultdict(int)
         self.asyncio = asyncio
+        self.function_and_status_codes_bundle: dict[str, dict[str, str]] = {}
+
+    def add_status_codes_to_bundle(
+        self, func_name: str, status_code_map: dict[str, str]
+    ) -> None:
+        """
+        Build a huge map of each function and it's status code responses.
+        At the end of the client generation you should call http_generator.generate_http_content()
+        """
+        self.function_and_status_codes_bundle[func_name] = status_code_map
+
+    def writeable_function_and_status_codes_bundle(self) -> str:
+        return f"func_response_code_maps = {self.function_and_status_codes_bundle}"
 
     def generate_http_content(self) -> None:
+        write_to_http(
+            self.writeable_function_and_status_codes_bundle(), self.output_dir
+        )
         client_generated = False
         client_type = "AsyncClient" if self.asyncio else "Client"
         if security_schemes := self.spec["components"].get("securitySchemes"):